renamed the remaining packages
