+++ .gitignore
@@ -32,5 +32,5 @@ target/
 logs/**
 logs
 build/
-customClass/
-customClass/**
+customClasses/
+customClasses/**
