+++ platform_bot/src/main/java/com/override/service/JoinRequestService.java
@@ -1,4 +1,4 @@
-package com.override.services;
+package com.override.service;
 
 import com.override.PlatformBot;
 import com.override.feign.NotificatorFeign;
