+++ platform_bot/src/main/java/com/override/service/RegisterService.java
@@ -1,4 +1,4 @@
-package com.override.services;
+package com.override.service;
 
 import com.override.feign.OrchestratorFeign;
 import dto.JoinRequestStatusDTO;
