+++ platform_code_executor/src/main/java/com/override/service/test/HelloWorldTaskTest.java
@@ -1,7 +1,7 @@
 package com.override.service.test;
 
 import com.override.exception.ExecutingCodeException;
-import com.override.utils.TestUtils;
+import com.override.util.TestUtils;
 import dto.TaskIdentifierDTO;
 import dto.TestResultDTO;
 import enums.CodeExecutionStatus;
