+++ platform_orchestrator/src/main/java/com/override/util/NavbarElement.java
@@ -2,7 +2,7 @@ package com.override.util;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.Getter;
+
 
 @Data
 @AllArgsConstructor
